quality: zero compiler/clippy warnings; alloc-free fast paths; single-allocation descriptors
- clippy 137→0 (auto-fixes + collapsible_match module allow with
  rationale + manual: iterator loops in merge_states/leaders/single,
  strip_prefix, sort_by_key, ptr_arg, match_single_binding,
  only_used_in_recursion with justification, too_many_arguments allow)
- rustc 3→0 (unreachable trailing expression, dead java_nested,
  unused binding)
- lift: the generation-aware final-read analysis now runs ONLY for
  blocks containing New/NewArray/FilledNewArray (the overwhelming
  majority skip the event walk entirely)
- PoolMethod descriptors build with ONE pre-sized String instead of
  params.join + format! (two intermediates × ~8M methods on lark)
- PendingMonitor type alias shared across classdec/browse/main;
  ResolvedTargets for prefix_or_full

57 tests green; weibo 3-run: 6.01-6.29s; lark 6.2s / 1.58GB RSS.

+++ crates/ddc-dex/src/reader.rs
@@ -165,7 +165,7 @@ impl<'a> Cursor<'a> {
                                 let lo = (((self.data[i + 3] & 0x0f) as u32) << 12)
                                     | (((self.data[i + 4] & 0x3f) as u32) << 6)
                                     | ((self.data[i + 5] & 0x3f) as u32);
-                                let cp = 0x10000 + (((c as u32) - 0xd800) << 10) + (lo - 0xdc00);
+                                let cp = 0x10000 + ((c - 0xd800) << 10) + (lo - 0xdc00);
                                 out.push(char::from_u32(cp).unwrap_or('\u{fffd}'));
                                 i += 6;
                                 units += 2;
